Bring medical branch up to date with webERP master branch

+++ CreditStatus.php
@@ -1,10 +1,11 @@
 <?php
-/* $Revision: 1.16 $ */
-/* $Id$*/
 
-include('includes/session.inc');
-$title = _('Credit Status Code Maintenance');
-include('includes/header.inc');
+
+include('includes/session.php');
+$Title = _('Credit Status Code Maintenance');
+$ViewTopic = 'CreditStatus';
+$BookMark = 'CreditStatus';
+include('includes/header.php');
 
 if (isset($_GET['SelectedReason'])){
 	$SelectedReason = $_GET['SelectedReason'];
@@ -17,7 +18,10 @@ if (isset($Errors)) {
 }
 $Errors = array();
 $InputError = 0;
-echo '<p class="page_title_text"><img src="'.$rootpath.'/css/'.$theme.'/images/maintenance.png" title="' . _('Search') . '" alt="" />' . ' ' . $title.'</p><br />';
+echo '<p class="page_title_text">
+		<img src="'.$RootPath.'/css/'.$Theme.'/images/maintenance.png" title="' . _('Search') . '" alt="" />' . ' ' . $Title.'
+	</p>
+	<br />';
 
 if (isset($_POST['submit'])) {
 
@@ -29,12 +33,12 @@ if (isset($_POST['submit'])) {
 
 	//first off validate inputs are sensible
 
-	$sql="SELECT count(reasoncode)
+	$SQL="SELECT count(reasoncode)
 			FROM holdreasons WHERE reasoncode='".$_POST['ReasonCode']."'";
-	$result=DB_query($sql, $db);
-	$myrow=DB_fetch_row($result);
+	$Result=DB_query($SQL);
+	$MyRow=DB_fetch_row($Result);
 
-	if ($myrow[0]!=0 and !isset($SelectedReason)) {
+	if ($MyRow[0]!=0 and !isset($SelectedReason)) {
 		$InputError = 1;
 		prnMsg( _('The credit status code already exists in the database'),'error');
 		$Errors[$i] = 'ReasonCode';
@@ -46,11 +50,11 @@ if (isset($_POST['submit'])) {
 		$Errors[$i] = 'ReasonCode';
 		$i++;
 	}
-	if (strlen($_POST['ReasonDescription']) > 30) {
+	if (mb_strlen($_POST['ReasonDescription']) > 30) {
 		$InputError = 1;
 		prnMsg(_('The credit status description must be thirty characters or less long'),'error');
 	}
-	if (strlen($_POST['ReasonDescription']) == 0) {
+	if (mb_strlen($_POST['ReasonDescription']) == 0) {
 		$InputError = 1;
 		prnMsg(_('The credit status description must be entered'),'error');
 		$Errors[$i] = 'ReasonDescription';
@@ -64,12 +68,12 @@ if (isset($_POST['submit'])) {
 		/*SelectedReason could also exist if submit had not been clicked this code would not run in this case cos submit is false of course	see the delete code below*/
 
 		if (isset($_POST['DisallowInvoices']) and $_POST['DisallowInvoices']=='on'){
-			$sql = "UPDATE holdreasons SET
+			$SQL = "UPDATE holdreasons SET
 							reasondescription='" . $_POST['ReasonDescription'] . "',
 							dissallowinvoices=1
 							WHERE reasoncode = '".$SelectedReason."'";
 		} else {
-			$sql = "UPDATE holdreasons SET
+			$SQL = "UPDATE holdreasons SET
 							reasondescription='" . $_POST['ReasonDescription'] . "',
 							dissallowinvoices=0
 							WHERE reasoncode = '".$SelectedReason."'";
@@ -80,54 +84,53 @@ if (isset($_POST['submit'])) {
 
 	/*Selected Reason is null cos no item selected on first time round so must be adding a record must be submitting new entries in the new status code form */
 
-		if (isset($_POST['DisallowInvoices']) and $_POST['DisallowInvoices']=='on'){
+		if (isset($_POST['DisallowInvoices']) AND $_POST['DisallowInvoices']=='on'){
 
-			$sql = "INSERT INTO holdreasons (
-					reasoncode,
-					reasondescription,
-					dissallowinvoices)
-					VALUES (
-					'" .$_POST['ReasonCode'] . "',
-					'".$_POST['ReasonDescription'] . "'
-					, 1)";
+			$SQL = "INSERT INTO holdreasons (reasoncode,
+											reasondescription,
+											dissallowinvoices)
+									VALUES ('" .$_POST['ReasonCode'] . "',
+											'".$_POST['ReasonDescription'] . "',
+											1)";
 		} else {
-			$sql = "INSERT INTO holdreasons (
-					reasoncode,
-					reasondescription,
-					dissallowinvoices)
-					VALUES (
-					'" . $_POST['ReasonCode'] . "',
-					'" . $_POST['ReasonDescription'] ."', 0)";
+			$SQL = "INSERT INTO holdreasons (reasoncode,
+											reasondescription,
+											dissallowinvoices)
+									VALUES ('" . $_POST['ReasonCode'] . "',
+											'" . $_POST['ReasonDescription'] ."',
+											0)";
 		}
 
 		$msg = _('A new credit status record has been inserted');
-		unset ($SelectedReason);
-		unset ($_POST['ReasonDescription']);
 	}
 	//run the SQL from either of the above possibilites
-	$result = DB_query($sql,$db);
+	$Result = DB_query($SQL);
 	if ($msg != '') {
 		prnMsg($msg,'success');
 	}
+	unset ($SelectedReason);
+	unset ($_POST['ReasonCode']);
+	unset ($_POST['ReasonDescription']);
+	unset ($_POST['submit']);
 } elseif (isset($_GET['delete'])) {
 //the link to delete a selected record was clicked instead of the submit button
 
 // PREVENT DELETES IF DEPENDENT RECORDS IN DebtorsMaster
 
-	$sql= "SELECT COUNT(*)
+	$SQL= "SELECT COUNT(*)
 			FROM debtorsmaster
 			WHERE debtorsmaster.holdreason='".$SelectedReason."'";
 
-	$result = DB_query($sql,$db);
-	$myrow = DB_fetch_row($result);
-	if ($myrow[0] > 0) {
+	$Result = DB_query($SQL);
+	$MyRow = DB_fetch_row($Result);
+	if ($MyRow[0] > 0) {
 		prnMsg( _('Cannot delete this credit status code because customer accounts have been created referring to it'),'warn');
-		echo '<br />' . _('There are') . ' ' . $myrow[0] . ' ' . _('customer accounts that refer to this credit status code');
+		echo '<br />' . _('There are') . ' ' . $MyRow[0] . ' ' . _('customer accounts that refer to this credit status code');
 	}  else {
 		//only delete if used in neither customer or supplier accounts
 
-		$sql="DELETE FROM holdreasons WHERE reasoncode=$SelectedReason";
-		$result = DB_query($sql,$db);
+		$SQL="DELETE FROM holdreasons WHERE reasoncode='" . $SelectedReason . "'";
+		$Result = DB_query($SQL);
 		prnMsg(_('This credit status code has been deleted'),'success');
 	}
 	//end if status code used in customer or supplier accounts
@@ -143,44 +146,38 @@ then none of the above are true and the list of status codes will be displayed w
 links to delete or edit each. These will call the same page again and allow update/input
 or deletion of the records*/
 
-	$sql = "SELECT reasoncode, reasondescription, dissallowinvoices FROM holdreasons";
-	$result = DB_query($sql, $db);
+	$SQL = "SELECT reasoncode, reasondescription, dissallowinvoices FROM holdreasons";
+	$Result = DB_query($SQL);
 
-	echo '<table class="selection">';
-	echo '<tr>
-		<th>'. _('Status Code') .'</th>
-		<th>'. _('Description') .'</th>
-		<th>'. _('Disallow Invoices') .'</th>';
+	echo '<table class="selection">
+		<tr>
+			<th>' .  _('Status Code')  . '</th>
+			<th>' .  _('Description')  . '</th>
+			<th>' .  _('Disallow Invoices')  . '</th>
+        </tr>';
 
-	$k=0; //row colour counter
-	while ($myrow=DB_fetch_row($result)) {
+	while ($MyRow=DB_fetch_array($Result)) {
 
-		if ($myrow[2]==0) {
+		if ($MyRow['dissallowinvoices']==0) {
 			$DissallowText = _('Invoice OK');
 		} else {
-			$DissallowText = '<b>'. _('NO INVOICING') .'</b>';
-		}
-		if ($k==1){
-			echo '<tr class="EvenTableRows">';
-			$k=0;
-		} else {
-			echo '<tr class="OddTableRows">';
-			$k=1;
+			$DissallowText = '<b>' .  _('NO INVOICING')  . '</b>';
 		}
 
-	printf('<td>%s</td>
-		<td>%s</td>
-		<td>%s</td>
-		<td><a href="%s?SelectedReason=%s">' . _('Edit') . '</a></td>
-		<td><a href="%s?SelectedReason=%s&delete=1">'. _('Delete') .'</a></td>
-		</tr>',
-		$myrow[0],
-		$myrow[1],
-		$DissallowText,
-		htmlspecialchars($_SERVER['PHP_SELF'], ENT_QUOTES, 'UTF-8'),
-		$myrow[0],
-		htmlspecialchars($_SERVER['PHP_SELF'], ENT_QUOTES, 'UTF-8'),
-		$myrow[0]);
+		printf('<tr class="striped_row">
+			<td>%s</td>
+			<td>%s</td>
+			<td>%s</td>
+			<td><a href="%s?SelectedReason=%s">' . _('Edit') . '</a></td>
+			<td><a href="%s?SelectedReason=%s&amp;delete=1" onclick="return confirm(\'' . _('Are you sure you wish to delete this credit status record?') . '\');">' .  _('Delete')  . '</a></td>
+			</tr>',
+			$MyRow['reasoncode'],
+			$MyRow['reasondescription'],
+			$DissallowText,
+			htmlspecialchars($_SERVER['PHP_SELF'],ENT_QUOTES,'UTF-8'),
+			$MyRow['reasoncode'],
+			htmlspecialchars($_SERVER['PHP_SELF'],ENT_QUOTES,'UTF-8'),
+			$MyRow['reasoncode']);
 
 	} //END WHILE LIST LOOP
 	echo '</table>';
@@ -188,60 +185,75 @@ or deletion of the records*/
 } //end of ifs and buts!
 
 if (isset($SelectedReason)) {
-	echo '<div class="centre"><a href="' . htmlspecialchars($_SERVER['PHP_SELF'], ENT_QUOTES, 'UTF-8') . '">' . _('Show Defined Credit Status Codes') . '</a></div>';
+	echo '<div class="centre">
+			<a href="' . htmlspecialchars($_SERVER['PHP_SELF'],ENT_QUOTES,'UTF-8') . '">' . _('Show Defined Credit Status Codes') . '</a>
+		</div>';
 }
 
 if (!isset($_GET['delete'])) {
 
-	echo '<form method="post" action="' . htmlspecialchars($_SERVER['PHP_SELF'], ENT_QUOTES, 'UTF-8') . '">';
+	echo '<form method="post" action="' . htmlspecialchars($_SERVER['PHP_SELF'],ENT_QUOTES,'UTF-8') . '">';
 	echo '<input type="hidden" name="FormID" value="' . $_SESSION['FormID'] . '" />';
 
 	if (isset($SelectedReason) and ($InputError!=1)) {
 		//editing an existing status code
 
-		$sql = "SELECT reasoncode,
-				reasondescription,
-				dissallowinvoices
-			FROM holdreasons
-			WHERE reasoncode='".$SelectedReason."'";
+		$SQL = "SELECT reasoncode,
+					reasondescription,
+					dissallowinvoices
+				FROM holdreasons
+				WHERE reasoncode='".$SelectedReason."'";
 
-		$result = DB_query($sql, $db);
-		$myrow = DB_fetch_array($result);
+		$Result = DB_query($SQL);
+		$MyRow = DB_fetch_array($Result);
 
-		$_POST['ReasonCode'] = $myrow['reasoncode'];
-		$_POST['ReasonDescription']  = $myrow['reasondescription'];
-		$_POST['DisallowInvoices']  = $myrow['dissallowinvoices'];
+		$_POST['ReasonCode'] = $MyRow['reasoncode'];
+		$_POST['ReasonDescription']  = $MyRow['reasondescription'];
+		$_POST['DisallowInvoices']  = $MyRow['dissallowinvoices'];
 
 		echo '<input type="hidden" name="SelectedReason" value="' . $SelectedReason . '" />';
 		echo '<input type="hidden" name="ReasonCode" value="' . $_POST['ReasonCode'] . '" />';
-		echo '<table class="selection"><tr><td>'. _('Status Code') .':</td><td>';
-		echo $_POST['ReasonCode'] . '</td></tr>';
+		echo '<fieldset>
+				<legend>', _('Edit Credit Status'), '</legend>
+				<field>
+					<label for="ReasonCode">' .  _('Status Code') .':</label>
+					<fieldtext>' . $_POST['ReasonCode'] . '</fieldtext>
+				</field>';
 
 	} else { //end of if $SelectedReason only do the else when a new record is being entered
 		if (!isset($_POST['ReasonCode'])) {
 			$_POST['ReasonCode'] = '';
 		}
-		echo '<br /><table class="selection">
-			<tr>
-				<td>'. _('Status Code') .':</td>
-				<td><input tabindex="1" type="text" name="ReasonCode" value="'. $_POST['ReasonCode'] .'" size="3" maxlength="2" /></td>
-			</tr>';
+		echo '<fieldset>
+				<legend>', _('Create Credit Status'), '</legend>
+				<field>
+					<label for="ReasonCode">' .  _('Status Code') .':</label>
+					<input ' . (in_array('ReasonCode',$Errors) ? 'class="integer inputerror"' : 'class="integer"' ) . ' tabindex="1" type="text" name="ReasonCode" required="required" value="'. $_POST['ReasonCode'] .'" size="3" maxlength="2" />
+				</field>';
 	}
 
 	if (!isset($_POST['ReasonDescription'])) {
 		$_POST['ReasonDescription'] = '';
 	}
-	echo '<tr>
-		<td>'. _('Description') .':</td>
-		<td><input tabindex="2" type="text" name="ReasonDescription" value="'. $_POST['ReasonDescription'] .'" size="28" maxlength="30" />
-		</td></tr>
-		<tr><td>'. _('Disallow Invoices') . '</td>';
+	echo '<field>
+			<label for="ReasonDescription">' .  _('Description') .':</label>
+			<input ' . (in_array('ReasonDescription',$Errors) ? 'class="inputerror"' : '' ) .
+			 ' tabindex="2" type="text" name="ReasonDescription" required="required" value="'. $_POST['ReasonDescription'] .'" size="28" maxlength="30" />
+		</field>
+		<field>
+			<label for="DisallowInvoices">' .  _('Disallow Invoices') . '</label>';
 	if (isset($_POST['DisallowInvoices']) and $_POST['DisallowInvoices']==1) {
-		echo '<td><input tabindex="3" type="checkbox" checked="True" name="DisallowInvoices" /></td></tr>';
+		echo '<input tabindex="3" type="checkbox" checked="checked" name="DisallowInvoices" />
+			</field>';
 	} else {
-		echo '<td><input tabindex="3" type="checkbox" name="DisallowInvoices" /></td></tr>';
+		echo '<input tabindex="3" type="checkbox" name="DisallowInvoices" />
+			</field>';
 	}
-	echo '</table><br /><div class="centre"><button tabindex="4" type="submit" name="submit" value="" />' . _('Enter Information') . '</button></div><br /></form>';
+	echo '</fieldset>
+			<div class="centre">
+				<input tabindex="4" type="submit" name="submit" value="' . _('Enter Information') . '" />
+            </div>
+			</form>';
 } //end if record deleted no point displaying form to add record
-include('includes/footer.inc');
+include('includes/footer.php');
 ?>
