+++ CustomerTypes.php
@@ -1,19 +1,15 @@
 <?php
-/* $Revision: 1.6 $ */
-/* $Id$*/
 
-include('includes/session.inc');
-$title = _('Customer Types') . ' / ' . _('Maintenance');
-include('includes/header.inc');
+include('includes/session.php');
+$Title = _('Customer Types') . ' / ' . _('Maintenance');
+$ViewTopic = 'Setup';
+$BookMark = 'CustomerTypes';
+include('includes/header.php');
 
 if (isset($_POST['SelectedType'])){
-	$SelectedType = strtoupper($_POST['SelectedType']);
+	$SelectedType = mb_strtoupper($_POST['SelectedType']);
 } elseif (isset($_GET['SelectedType'])){
-	$SelectedType = strtoupper($_GET['SelectedType']);
-}
-
-if (isset($_GET['name'])) {
-	$_POST['typename']=$_GET['name'];
+	$SelectedType = mb_strtoupper($_GET['SelectedType']);
 }
 
 if (isset($Errors)) {
@@ -22,7 +18,8 @@ if (isset($Errors)) {
 
 $Errors = array();
 
-echo '<p class="page_title_text"><img src="'.$rootpath.'/css/'.$theme.'/images/maintenance.png" title="' . _('Customer Types') . '" alt="" />' . _('Customer Type Setup') . '</p>';
+echo '<p class="page_title_text"><img src="'.$RootPath.'/css/'.$Theme.'/images/maintenance.png" title="' . _('Customer Types') .
+	'" alt="" />' . _('Customer Type Setup') . '</p>';
 echo '<div class="page_help_text">' . _('Add/edit/delete Customer Types') . '</div>';
 
 if (isset($_POST['submit'])) {
@@ -35,14 +32,14 @@ if (isset($_POST['submit'])) {
 
 	//first off validate inputs sensible
 	$i=1;
-	if (strlen($_POST['typename']) >100) {
+	if (mb_strlen($_POST['TypeName']) >100) {
 		$InputError = 1;
 		prnMsg(_('The customer type name description must be 100 characters or less long'),'error');
 		$Errors[$i] = 'CustomerType';
 		$i++;
 	}
 
-	if (strlen($_POST['typename'])==0) {
+	if (mb_strlen($_POST['TypeName'])==0) {
 		$InputError = 1;
 		echo '<br />';
 		prnMsg(_('The customer type name description must contain at least one character'),'error');
@@ -52,13 +49,13 @@ if (isset($_POST['submit'])) {
 
 	$checksql = "SELECT count(*)
 		     FROM debtortype
-		     WHERE typename = '" . $_POST['typename'] . "'";
-	$checkresult=DB_query($checksql, $db);
+		     WHERE typename = '" . $_POST['TypeName'] . "'";
+	$checkresult=DB_query($checksql);
 	$checkrow=DB_fetch_row($checkresult);
 	if ($checkrow[0]>0 and !isset($SelectedType)) {
 		$InputError = 1;
 		echo '<br />';
-		prnMsg(_('You already have a customer type called').' '.$_POST['typename'],'error');
+		prnMsg(_('You already have a customer type called').' '.$_POST['TypeName'],'error');
 		$Errors[$i] = 'CustomerName';
 		$i++;
 	}
@@ -66,37 +63,37 @@ if (isset($_POST['submit'])) {
 	if (isset($SelectedType) AND $InputError !=1) {
 
 		$sql = "UPDATE debtortype
-			SET typename = '" . $_POST['typename'] . "'
+			SET typename = '" . $_POST['TypeName'] . "'
 			WHERE typeid = '" .$SelectedType."'";
 
-		$msg = _('The customer type') . ' ' . $_POST['typename'] . ' ' .  _('has been updated');
+		$msg = _('The customer type') . ' ' . $SelectedType . ' ' .  _('has been updated');
 	} elseif ( $InputError !=1 ) {
 
 		// First check the type is not being duplicated
 
 		$checkSql = "SELECT count(*)
 			     FROM debtortype
-			     WHERE typename = '" . $_POST['typename'] . "'";
+			     WHERE typename = '" . $_POST['TypeName'] . "'";
 
-		$checkresult = DB_query($checkSql,$db);
+		$checkresult = DB_query($checkSql);
 		$checkrow = DB_fetch_row($checkresult);
 
 		if ( $checkrow[0] > 0 ) {
 			$InputError = 1;
-			prnMsg( _('The customer type ') . $_POST['typeid'] . _(' already exist.'),'error');
+			prnMsg( _('The customer type') . ' ' . $_POST['typeid'] . _(' already exist.'),'error');
 		} else {
 
 			// Add new record on submit
 
 			$sql = "INSERT INTO debtortype
 						(typename)
-					VALUES ('" . $_POST['typename'] . "')";
+					VALUES ('" . $_POST['TypeName'] . "')";
 
 
-			$msg = _('Customer type') . ' ' . $_POST['typename'] .  ' ' . _('has been created');
+			$msg = _('Customer type') . ' ' . $_POST["typename"] .  ' ' . _('has been created');
 			$checkSql = "SELECT count(typeid)
 			     FROM debtortype";
-			$result = DB_query($checkSql, $db);
+			$result = DB_query($checkSql);
 			$row = DB_fetch_row($result);
 
 		}
@@ -104,7 +101,7 @@ if (isset($_POST['submit'])) {
 
 	if ( $InputError !=1) {
 	//run the SQL from either of the above possibilites
-		$result = DB_query($sql,$db);
+		$result = DB_query($sql);
 
 
 	// Fetch the default price list.
@@ -114,7 +111,7 @@ if (isset($_POST['submit'])) {
 		$checkSql = "SELECT count(*)
 			     FROM debtortype
 			     WHERE typeid = '" . $DefaultCustomerType . "'";
-		$checkresult = DB_query($checkSql,$db);
+		$checkresult = DB_query($checkSql);
 		$checkrow = DB_fetch_row($checkresult);
 
 	// If it doesnt then update config with newly created one.
@@ -122,7 +119,7 @@ if (isset($_POST['submit'])) {
 			$sql = "UPDATE config
 					SET confvalue='" . $_POST['typeid'] . "'
 					WHERE confname='DefaultCustomerType'";
-			$result = DB_query($sql,$db);
+			$result = DB_query($sql);
 			$_SESSION['DefaultCustomerType'] = $_POST['typeid'];
 		}
 		echo '<br />';
@@ -130,7 +127,7 @@ if (isset($_POST['submit'])) {
 
 		unset($SelectedType);
 		unset($_POST['typeid']);
-		unset($_POST['typename']);
+		unset($_POST['TypeName']);
 	}
 
 } elseif ( isset($_GET['delete']) ) {
@@ -143,7 +140,7 @@ if (isset($_POST['submit'])) {
 	       WHERE debtortrans.type='".$SelectedType."'";
 
 	$ErrMsg = _('The number of transactions using this customer type could not be retrieved');
-	$result = DB_query($sql,$db,$ErrMsg);
+	$result = DB_query($sql,$ErrMsg);
 
 	$myrow = DB_fetch_row($result);
 	if ($myrow[0]>0) {
@@ -154,18 +151,22 @@ if (isset($_POST['submit'])) {
 		$sql = "SELECT COUNT(*) FROM debtorsmaster WHERE typeid='".$SelectedType."'";
 
 		$ErrMsg = _('The number of transactions using this Type record could not be retrieved because');
-		$result = DB_query($sql,$db,$ErrMsg);
+		$result = DB_query($sql,$ErrMsg);
 		$myrow = DB_fetch_row($result);
 		if ($myrow[0]>0) {
 			prnMsg (_('Cannot delete this type because customers are currently set up to use this type') . '<br />' . _('There are') . ' ' . $myrow[0] . ' ' . _('customers with this type code'));
 		} else {
+			$result = DB_query("SELECT typename FROM debtortype WHERE typeid='".$SelectedType."'");
+			if (DB_Num_Rows($result)>0){
+				$TypeRow = DB_fetch_array($result);
+				$TypeName = $TypeRow['typename'];
 
-			$sql="DELETE FROM debtortype WHERE typeid='".$SelectedType."'";
-			$ErrMsg = _('The Type record could not be deleted because');
-			$result = DB_query($sql,$db,$ErrMsg);
-			echo '<br />';
-			prnMsg(_('Customer type') . ' ' . $_POST['typename']  . ' ' . _('has been deleted') ,'success');
-
+				$sql="DELETE FROM debtortype WHERE typeid='".$SelectedType."'";
+				$ErrMsg = _('The Type record could not be deleted because');
+				$result = DB_query($sql,$ErrMsg);
+				echo '<br />';
+				prnMsg(_('Customer type') . ' ' . $TypeName  . ' ' . _('has been deleted') ,'success');
+			}
 			unset ($SelectedType);
 			unset($_GET['delete']);
 
@@ -181,99 +182,93 @@ links to delete or edit each. These will call the same page again and allow upda
 or deletion of the records*/
 
 	$sql = "SELECT typeid, typename FROM debtortype";
-	$result = DB_query($sql,$db);
+	$result = DB_query($sql);
 
-	echo '<br /><table class="selection">';
-	echo '<tr>
-		<th>' . _('Type ID') . '</th>
-		<th>' . _('Type Name') . '</th>
-		</tr>';
+	echo '<table class="selection">';
+	echo '<thead>
+			<tr>
+			<th class="ascending">' . _('Type ID') . '</th>
+			<th class="ascending">' . _('Type Name') . '</th>
+			</tr>
+		</thead>
+		<tbody>';
 
-$k=0; //row colour counter
+while ($myrow = DB_fetch_row($result)) {
 
-while ($myrow = DB_fetch_array($result)) {
-	if ($k==1){
-		echo '<tr class="EvenTableRows">';
-		$k=0;
-	} else {
-		echo '<tr class="OddTableRows">';
-		$k=1;
-	}
-
-	printf('
+	printf('<tr class="striped_row">
 		<td>%s</td>
 		<td>%s</td>
-		<td><a href="%sSelectedType=%s">' . _('Edit') . '</td>
-		<td><a href="%sSelectedType=%s&delete=yes&name=%s" onclick=\'return confirm("' . _('Are you sure you wish to delete this Customer Type?') . '");\'>' . _('Delete') . '</td>
+		<td><a href="%sSelectedType=%s">' . _('Edit') . '</a></td>
+		<td><a href="%sSelectedType=%s&amp;delete=yes" onclick=\'return confirm("' . _('Are you sure you wish to delete this Customer Type?') . '");\'>' . _('Delete') . '</a></td>
 		</tr>',
-		$myrow['typeid'],
-		$myrow['typename'],
-		htmlspecialchars($_SERVER['PHP_SELF'], ENT_QUOTES, 'UTF-8') . '?',
-		$myrow['typeid'],
-		htmlspecialchars($_SERVER['PHP_SELF'], ENT_QUOTES, 'UTF-8') . '?',
-		$myrow['typeid'],
-		$myrow['typename']);
+		$myrow[0],
+		$myrow[1],
+		htmlspecialchars($_SERVER['PHP_SELF'],ENT_QUOTES,'UTF-8') . '?',
+		$myrow[0],
+		htmlspecialchars($_SERVER['PHP_SELF'],ENT_QUOTES,'UTF-8') . '?',
+		$myrow[0]);
 	}
 	//END WHILE LIST LOOP
-	echo '</table>';
+	echo '</tbody></table>';
 }
 
 //end of ifs and buts!
 if (isset($SelectedType)) {
 
-	echo '<div class="centre"><p><a href="' . htmlspecialchars($_SERVER['PHP_SELF'], ENT_QUOTES, 'UTF-8') . '">' . _('Show All Types Defined') . '</a></p></div>';
+	echo '<div class="centre"><br /><a href="' . htmlspecialchars($_SERVER['PHP_SELF'],ENT_QUOTES,'UTF-8') . '">' . _('Show All Types Defined') . '</a></div>';
 }
 if (! isset($_GET['delete'])) {
 
-	echo '<form method="post" action="' . htmlspecialchars($_SERVER['PHP_SELF'], ENT_QUOTES, 'UTF-8') . '">';
-	echo '<input type="hidden" name="FormID" value="' . $_SESSION['FormID'] . '" />';
-	echo '<table class="selection">'; //Main table
-
+	echo '<form method="post" action="' . htmlspecialchars($_SERVER['PHP_SELF'],ENT_QUOTES,'UTF-8') .  '">
+		<input type="hidden" name="FormID" value="' . $_SESSION['FormID'] . '" />';
 
 	// The user wish to EDIT an existing type
-	if ( isset($SelectedType) AND $SelectedType!='' )
-	{
+	if ( isset($SelectedType) AND $SelectedType!='' ) {
 
 		$sql = "SELECT typeid,
 			       typename
 		        FROM debtortype
 		        WHERE typeid='".$SelectedType."'";
 
-		$result = DB_query($sql, $db);
+		$result = DB_query($sql);
 		$myrow = DB_fetch_array($result);
 
 		$_POST['typeid'] = $myrow['typeid'];
-		$_POST['typename']  = $myrow['typename'];
+		$_POST['TypeName']  = $myrow['typename'];
 
-		echo '<input type="hidden" name="SelectedType" value="' . $SelectedType . '" />';
-		echo '<input type="hidden" name="typeid" value="' . $_POST['typeid'] . '" />';
-		echo '<table class="selection">';
+		echo '<input type="hidden" name="SelectedType" value="' . $SelectedType . '" />
+			<input type="hidden" name="typeid" value="' . $_POST['typeid'] . '" />';
+
+		echo '<fieldset>
+				<legend>', _('Edit Customer Type'), '</legend>';
 
 		// We dont allow the user to change an existing type code
 
-		echo '<tr><td>' . _('Type ID') . ': ' . $_POST['typeid'] . '</td></tr>';
-
+		echo '<field>
+				<label for="typeid">' . _('Type ID') . ':</label>
+				<fieldtext>' . $_POST['typeid'] . '</fieldtext>
+			</field>';
 	} else 	{
-
 		// This is a new type so the user may volunteer a type code
-
-		echo '<table class="selection">';
-
+		echo '<fieldset>
+				<legend>', _('Create New Customer Type'), '</legend>';
 	}
 
-	if (!isset($_POST['typename'])) {
-		$_POST['typename']='';
+	if (!isset($_POST['TypeName'])) {
+		$_POST['TypeName']='';
 	}
-	echo '<tr><td>' . _('Type Name') . ':</td><td><input type="text" name="typename" value="' . $_POST['typename'] . '" /></td></tr>';
-
-   	echo '</td></tr></table>'; // close main table
-
-	echo '<br /><div class="centre"><button type="submit" name="submit">' . _('Accept') . '</button></div><br />';
-
-	echo '</form>';
+	echo '<field>
+			<label for="TypeName">' . _('Type Name') . ':</label>
+			<input type="text" name="TypeName"  required="required" title="" value="' . $_POST['TypeName'] . '" />
+			<fieldhelp>' . _('The customer type name is required') . '</fieldhelp
+		</field>
+		</fieldset>
+		<div class="centre">
+			<input type="submit" name="submit" value="' . _('Accept') . '" />
+		</div>
+	</form>';
 
 } // end if user wish to delete
 
-
-include('includes/footer.inc');
+include('includes/footer.php');
 ?>
